+++ examples/benchmark.py
@@ -7,7 +7,12 @@ import matplotlib.pyplot as plt
 from time import perf_counter
 
 
-def main(nbenches: int = 1, theta: float = 0.5, mesh_size_max: float = 0.033, mesh_size_min: float = 0.033):
+def main(
+    nbenches: int = 1,
+    theta: float = 0.5,
+    mesh_size_max: float = 0.033,
+    mesh_size_min: float = 0.033,
+):
     mesh_sizes = np.linspace(mesh_size_max, mesh_size_min, nbenches)
     direct_times = []
     direct_interactions = []
@@ -18,7 +23,9 @@ def main(nbenches: int = 1, theta: float = 0.5, mesh_size_max: float = 0.033, me
     interactions = np.zeros(nbenches)
 
     for i, mesh_size in enumerate(mesh_sizes):
-        mesh, jdensity = oersted.testing.make_helmholtz("tests/data/ring.stp", mesh_size)
+        mesh, jdensity = oersted.testing.make_helmholtz(
+            "tests/data/ring.stp", mesh_size
+        )
         n = mesh.num_elems
         interactions[i] = n * n
 
@@ -32,13 +39,17 @@ def main(nbenches: int = 1, theta: float = 0.5, mesh_size_max: float = 0.033, me
             est_direct_interactions = [n * n]
             i_est = i
         else:
-            m = (direct_times[i_est] - direct_times[0]) / (interactions[i_est] - interactions[0])
+            m = (direct_times[i_est] - direct_times[0]) / (
+                interactions[i_est] - interactions[0]
+            )
             b = direct_times[i_est] - m * interactions[i_est]
             est_direct_interactions.append(n * n)
             est_direct_times.append(m * n * n + b)
 
         start = perf_counter()
-        _ = oersted.b_field(mesh.to_centroid_mesh(), jdensity, mesh.centroids, OctreeSolver(theta=theta))
+        _ = oersted.b_field(
+            mesh.to_centroid_mesh(), jdensity, mesh.centroids, OctreeSolver(theta=theta)
+        )
         end = perf_counter()
         octree_times[i] = end - start
 
@@ -48,7 +59,10 @@ def main(nbenches: int = 1, theta: float = 0.5, mesh_size_max: float = 0.033, me
         n = np.sqrt(interactions[i])
         speedup = all_direct_times[i] / octree_times[i]
 
-        print(f"{n:.0f} | {interactions[i]:.3e} | {all_direct_times[i]:.3f}" + f"| {octree_times[i]:.3f} | {speedup:.3f}x")
+        print(
+            f"{n:.0f} | {interactions[i]:.3e} | {all_direct_times[i]:.3f}"
+            + f"| {octree_times[i]:.3f} | {speedup:.3f}x"
+        )
 
     # Plot solution times
     fig, ax = plt.subplots()
@@ -64,13 +78,22 @@ def main(nbenches: int = 1, theta: float = 0.5, mesh_size_max: float = 0.033, me
     fig.savefig("tests/fig/benchmarks.svg")
 
     # Plot interactions per second
-    direct_throughput = [i / (t * 1e9) for (i, t) in zip(direct_interactions, direct_times, strict=True)]
-    est_direct_throughput = [i / (t * 1e9) for (i, t) in zip(est_direct_interactions, est_direct_times, strict=True)]
-    octree_throughput = [i / (t * 1e9) for (i, t) in zip(interactions, octree_times, strict=True)]
+    direct_throughput = [
+        i / (t * 1e9) for (i, t) in zip(direct_interactions, direct_times, strict=True)
+    ]
+    est_direct_throughput = [
+        i / (t * 1e9)
+        for (i, t) in zip(est_direct_interactions, est_direct_times, strict=True)
+    ]
+    octree_throughput = [
+        i / (t * 1e9) for (i, t) in zip(interactions, octree_times, strict=True)
+    ]
 
     fig, ax = plt.subplots()
     ax.plot(direct_interactions, direct_throughput, "r", label="direct")
-    ax.plot(est_direct_interactions, est_direct_throughput, "r--", label="direct (trend)")
+    ax.plot(
+        est_direct_interactions, est_direct_throughput, "r--", label="direct (trend)"
+    )
     ax.plot(interactions, octree_throughput, "k", label="octree")
     ax.set_xlabel("Interactions ($N^2$)")
     ax.set_ylabel("Throughput ($1e9$ interactions/sec)")