+++ src/oersted/__init__.py
@@ -6,9 +6,15 @@ from .solver import Solver, DirectSolver, OctreeSolver
 from .biotsavart import b_field, h_field, h_mag
 
 from .mesh import Mesh, CentroidMesh, mesh_step, plot_mesh
-from . import testing
-from . import magnetization
-from .results import kelvin_force_density, kelvin_forces, maxwell_forces
+from .testing import make_helmholtz, smape
+from .magnetization import demag_solve
+from .results import (
+    kelvin_force_density,
+    kelvin_forces,
+    maxwell_forces,
+    lorentz_forces,
+    lorentz_force_density,
+)
 
 
 __all__ = [
@@ -17,7 +23,6 @@ __all__ = [
     "LinearMaterial",
     "NonlinearMaterial",
     "BHCurve",
-    "testing",
     "Mesh",
     "CentroidMesh",
     "mesh_step",
@@ -32,8 +37,12 @@ __all__ = [
     "Solver",
     "DirectSolver",
     "OctreeSolver",
-    "magnetization",
+    "demag_solve",
     "maxwell_forces",
     "kelvin_forces",
     "kelvin_force_density",
+    "lorentz_forces",
+    "lorentz_force_density",
+    "make_helmholtz",
+    "smape",
 ]