+++ src/oersted/magnetization.py
@@ -10,24 +10,27 @@ from .materials import Material
 from ._oersted import magnetization_tet4
 
 
-def demag_tet4(
-    mesh: Mesh, material: Material, h_external: NDArray[float64], solver: DirectSolver | OctreeSolver
+def demag_solve(
+    mesh: Mesh,
+    material: Material,
+    h_external: NDArray[float64],
+    solver: DirectSolver | OctreeSolver,
 ) -> tuple[NDArray[float64], NDArray[float64]]:
-    """Compute magnetization field M and the total H field at element centroids, given a background field
+    """Compute magnetization field M and the total H field at element centroids,
+        given a background field
 
-    Uses simple fixed-point iteration and therefore only converges for low-permeable materials.
+    Uses simple fixed-point iteration and therefore only converges for
+        low-permeable materials.
 
     Args:
-        nodes: (Nn, 3) nodal coordinates
-        element_connectivity: (Ne, 4) indices of each node per element;
-            these are indices of the array `nodes`, not of the solver's node numbers
+        mesh: finite element mesh on which to evaluate the demagnetizing field
         material: linear or nonlinear magnetic maaterial properties
-        h_external: (Ne,3) external field at each element centroid
+        h_external: (A/m) an (Ne,3) array of external field at each element centroid
         solver: solution parameters for the problem, including iteration method
 
     Returns:
-        (M, Htotal): each (Ne, 3), magnetization field M(Htotal) and total H field at element
-            centroids. These can be summed to give B = mu0 * (Htotal + M).
+        (M, Htotal): each (Ne, 3), magnetization field M(Htotal) and total H field
+        at element centroids. These can be summed to give B = mu0 * (Htotal + M).
     """
 
     theta: float