+++ src/oersted/solver.py
@@ -1,4 +1,10 @@
+"""Solution controls"""
+
+
 class Solver:
+    """Defines a generic solver with options for iterative and multi-threaded
+    solutions"""
+
     _n_threads: int
     _max_iterations: int
     _tol: float
@@ -19,7 +25,8 @@ class Solver:
         """Convergence tolerance for iterative solves.
 
         Note: this is typically a value of the H-field for magnetization calculations;
-        this is roughly 1.3e-6 T, which may be more than necessary for most applications.
+            the default value is roughly 1.3e-6 T, which may be more than necessary
+            for most applications.
         """
         return self._tol
 
@@ -30,6 +37,9 @@ class Solver:
 
 
 class DirectSolver(Solver):
+    """Controls solver options for using the direct (full integration)
+    solution routines"""
+
     def __init__(self, n_threads: int = 0, max_iterations=100, tol=1.0, alpha=0.5):
         self._n_threads = n_threads
         self._max_iterations = max_iterations
@@ -38,11 +48,20 @@ class DirectSolver(Solver):
 
 
 class OctreeSolver(Solver):
+    """Controls solver options for using the octree (Barnes-Hut) solution routines"""
+
     _theta: float
     _leaf_threshold: int
 
-    def __init__(self, theta: float = 0.25, leaf_threshold: int = 16, n_threads: int = 0, max_iterations=100, tol=1.0, alpha=0.5):
-
+    def __init__(
+        self,
+        theta: float = 0.25,
+        leaf_threshold: int = 16,
+        n_threads: int = 0,
+        max_iterations=100,
+        tol=1.0,
+        alpha=0.5,
+    ):
         self._theta = theta
         self._leaf_threshold = leaf_threshold
         self._n_threads = n_threads
@@ -52,8 +71,11 @@ class OctreeSolver(Solver):
 
     @property
     def theta(self):
+        """Returns the Barnes-Hut angle-opening criteria (accuracy control)"""
         return self._theta
 
     @property
     def leaf_threshold(self):
+        """Returns the number of sources that will be evaluated individually at the
+        octree leaf level"""
         return self._leaf_threshold