+++ tests/test_helmholtz.py
@@ -27,14 +27,16 @@ def setup_test():
 
     # Runtime parameters
     theta: float = 0.5
-    mesh_size: float = 0.010  # ~10M interactions; set to 33 for 1e6 interactions
+    mesh_size: float = 0.012  # ~10M interactions; set to 33 for 1e6 interactions
     ntargets_axis: int = 100  # Along the axis
     nthreads = 0
     leaf_threshold = 16
     axis_halfdistance = 0.01
 
     direct_solver = DirectSolver(n_threads=nthreads)
-    octree_solver = OctreeSolver(theta=theta, leaf_threshold=leaf_threshold, n_threads=nthreads)
+    octree_solver = OctreeSolver(
+        theta=theta, leaf_threshold=leaf_threshold, n_threads=nthreads
+    )
 
     #
     # Generate a mesh from a STEP file
@@ -44,7 +46,9 @@ def setup_test():
 
     # Setup the targets for the axis accuracy test
     targets_axis = np.zeros((ntargets_axis, 3))
-    targets_axis[:, 2] = np.linspace(-axis_halfdistance, axis_halfdistance, ntargets_axis)
+    targets_axis[:, 2] = np.linspace(
+        -axis_halfdistance, axis_halfdistance, ntargets_axis
+    )
 
     return mesh, jdensity, targets_axis, direct_solver, octree_solver
 
@@ -56,8 +60,12 @@ def rel_field_on_axis(mesh: Mesh, jdensity, targets_axis, direct_solver, octree_
 
     b_tet4_direct = oersted.b_field(mesh, jdensity, targets_axis, solver=direct_solver)
     b_tet4_octree = oersted.b_field(mesh, jdensity, targets_axis, solver=octree_solver)
-    b_point_direct = oersted.b_field(centroid_mesh, jdensity, targets_axis, solver=direct_solver)
-    b_point_octree = oersted.b_field(centroid_mesh, jdensity, targets_axis, solver=octree_solver)
+    b_point_direct = oersted.b_field(
+        centroid_mesh, jdensity, targets_axis, solver=direct_solver
+    )
+    b_point_octree = oersted.b_field(
+        centroid_mesh, jdensity, targets_axis, solver=octree_solver
+    )
 
     # Use the tet4 direct field as the comparison
     # Field should only be in the z-axis
@@ -79,10 +87,18 @@ def rel_field_on_axis(mesh: Mesh, jdensity, targets_axis, direct_solver, octree_
     current: float = 100e3
     bz_analytical = (0.8**1.5) * oersted.MU0 * current / 0.2
     target_center = np.array([[0.0, 0.0, 0.0]])
-    bz_tet4_direct = oersted.b_field(mesh, jdensity, target_center, solver=direct_solver)[0, 2]
-    bz_tet4_octree = oersted.b_field(mesh, jdensity, target_center, solver=octree_solver)[0, 2]
-    bz_point_direct = oersted.b_field(centroid_mesh, jdensity, target_center, solver=direct_solver)[0, 2]
-    bz_point_octree = oersted.b_field(centroid_mesh, jdensity, target_center, solver=octree_solver)[0, 2]
+    bz_tet4_direct = oersted.b_field(
+        mesh, jdensity, target_center, solver=direct_solver
+    )[0, 2]
+    bz_tet4_octree = oersted.b_field(
+        mesh, jdensity, target_center, solver=octree_solver
+    )[0, 2]
+    bz_point_direct = oersted.b_field(
+        centroid_mesh, jdensity, target_center, solver=direct_solver
+    )[0, 2]
+    bz_point_octree = oersted.b_field(
+        centroid_mesh, jdensity, target_center, solver=octree_solver
+    )[0, 2]
 
     assert np.abs(bz_analytical - bz_tet4_direct) / bz_analytical < 1e-3
     assert np.abs(bz_analytical - bz_tet4_octree) / bz_analytical < 1e-2